+++ src/auth/auth.service.ts
@@ -2,12 +2,15 @@ import { Injectable } from '@nestjs/common';
 import { UsersService } from '../users/users.service';
 import { JwtService } from '@nestjs/jwt';
 import { User } from 'src/users/user.entity';
+import { ProfilesService } from 'src/files/files.service';
+import { Profile } from 'src/files/file.entity';
 
 @Injectable()
 export class AuthService {
   constructor(
     private usersService: UsersService,
     private jwtService: JwtService,
+    private profileService: ProfilesService,
   ) {}
 
   async validateUser(username: string, pass: string): Promise<any> {
@@ -29,5 +32,10 @@ export class AuthService {
   async signup(user: User) {
     const new_user = await this.usersService.create(user);
     return user.email;
+  }
+
+  async updateProfile(profile: Profile) {
+    const new_user = await this.profileService.createProfile(profile);
+    return profile;
   }
 }
