+++ src/files/files.module.ts
@@ -1,0 +1,11 @@
+import { Module } from '@nestjs/common';
+import { TypeOrmModule } from '@nestjs/typeorm';
+import { File } from './file.entity';
+import { FilesService } from './files.service';
+
+@Module({
+  imports: [TypeOrmModule.forFeature([File])],
+  providers: [FilesService],
+  exports: [FilesService, TypeOrmModule],
+})
+export class ProfileModule {}
