+++ src/helpers/image-storage.ts
@@ -6,7 +6,6 @@ export const saveImageToStorage = {
   storage: diskStorage({
     destination: './storage/uploads',
     filename: (req, file, cb) => {
-      console.log(file);
       const fileExtension: string = path.extname(file.originalname);
       const fileName: string = uuid() + fileExtension;
       cb(null, fileName);
