--- src/profiles/profiles.module.ts
@@ -1,11 +1,0 @@
-import { Module } from '@nestjs/common';
-import { TypeOrmModule } from '@nestjs/typeorm';
-import { Profile } from '../profiles/profile.entity';
-import { ProfilesService } from './profiles.service';
-
-@Module({
-  imports: [TypeOrmModule.forFeature([Profile])],
-  providers: [ProfilesService],
-  exports: [TypeOrmModule],
-})
-export class ProfileModule {}
