+++ src/users/users.module.ts
@@ -2,13 +2,9 @@ import { Module } from '@nestjs/common';
 import { UsersService } from './users.service';
 import { User } from './user.entity';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { Profile } from '../profiles/profile.entity';
 
 @Module({
-  imports: [
-    TypeOrmModule.forFeature([User]),
-    TypeOrmModule.forFeature([Profile]),
-  ],
+  imports: [TypeOrmModule.forFeature([User])],
   providers: [UsersService],
   exports: [UsersService, TypeOrmModule],
 })
